used deepcopy to avoid copying states, new keys
deepcopy is now used to copy camera states, as simple copy was messing them up. New keys allow to replace and delete key frames. The key-frame capture key f now also inserts key-frames when one is in the middle of a series of them after moving with a or b.